Redirect implementation is moved to OAuth2 plugin. Internal resource class is removed. Some python2 capabilities are added. Resource of index page has got an optional anonymous restriction.
